make folder differentiation
