+++ horqrux/differentiation/__init__.py
@@ -1,0 +1,3 @@
+from __future__ import annotations
+
+from .automatic_diff import ad_expectation
